{{cookiecutter.project_name}}/{{cookiecutter.project_slug}}/app/utils/errors.py
from dataclasses import asdict

from drf_standardized_errors.formatter import ExceptionFormatter
from drf_standardized_errors.types import ErrorResponse
from rest_framework.exceptions import APIException


class CustomExceptionFormatter(ExceptionFormatter):
    def format_error_response(self, error_response: ErrorResponse):
        return {"status": "error", "type": error_response.type, "errors": [asdict(o) for o in error_response.errors]}


# example of custom exception
class ServiceUnavailable(APIException):
    status_code = 503
    default_detail = "Service temporarily unavailable, try again later."
    default_code = "service_unavailable"
