{{cookiecutter.project_name}}/requirements.txt
# Django
# ------------------------------------------------------------------------------
django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0  # https://github.com/adamchainz/django-cors-headers
gunicorn==21.2.0
drf-spectacular==0.26.5  # https://github.com/tfranzel/drf-spectacular
drf-standardized-errors==0.12.5  # https://github.com/ghazi-git/drf-standardized-errors
dapr==1.11.0  # https://github.com/dapr/python-sdk

# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0  # https://github.com/PyCQA/flake8
flake8-isort==6.1.0  # https://github.com/gforcada/flake8-isort
coverage==7.3.1  # https://github.com/nedbat/coveragepy
black==23.9.1  # https://github.com/psf/black
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pre-commit==3.4.0  # https://github.com/pre-commit/pre-commit
